Fix the build on FreeBSD (and OpenBSD)
For now, we use Dispatch for AsyncIO and process termination monitoring, since BSDs (including macOS) all use kqueue as the Dispatch backend anyways.

Closes #115
